Rename pybind_boost_header.h
