+++ paddle/fluid/pybind/io.h
@@ -21,7 +21,7 @@ typedef SSIZE_T ssize_t;
 
 #include <Python.h>
 
-#include "paddle/fluid/pybind/pybind_boost_headers.h"
+#include "paddle/fluid/pybind/pybind_variant_caster.h"
 
 namespace paddle {
 namespace pybind {
